feat: LottoChecker 분리, print 메서드 LottoChecker로 이동
