+++ src/constants.rs
@@ -1,0 +1,4 @@
+pub const MIN_NUMBER: u32 = 1;
+pub const MAX_NUMBER: u32 = 45;
+pub const LOTTO_LENGTH: u32 = 6;
+pub const LOTTO_PRICE: u32 = 1000;
